Feat: update endpoint
