+++ challenge_05/src/app.js
@@ -2,6 +2,8 @@ import express from "express";
 import {engine} from "express-handlebars";
 import __dirname from "./utils.js";
 import router from "./routers/index.js";
+import errorHandler from "./middlewares/errorHandler.js";
+import notFoundHandlers from "./middlewares/notFoundHandlers.js";
 
 const server = express();
 
@@ -17,5 +19,7 @@ server.set("views", __dirname + "/views");
 
 //Endpoints
 server.use("/", router);
+server.use(errorHandler);
+server.use(notFoundHandlers);
 
 export default server;
