+++ challenge_05/src/middlewares/multer.js
@@ -1,11 +1,20 @@
-import multer from 'multer'
-import { __dirname } from '../utils.js'
-
+import multer from 'multer';
+import __dirname from '../utils.js';
 const storage = multer.diskStorage({
-    destination: (req,file,cb) => cb(null,__dirname+'/public/img'),
-    filename: (req,file,cb) => cb(null,file.originalname)
+    destination:(req,file,cb)=>{
+        if(file.fieldname==="image"){
+            cb(null,__dirname+'/public/images');
+        }else if(file.fieldname==="documents"){
+            cb(null,'documents');
+        }
+        else{
+            cb(null,'uploads');
+        }
+    },
+    filename:(req,file,cb)=>{
+        cb(null,Date.now()+file.originalname);
+    }
 })
+const upload = multer({storage:storage});
 
-const uploader = multer({storage});
-
-export default uploader;
+export default upload;
